+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vjean <vjean@student.42quebec.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/10 10:22:49 by vjean             #+#    #+#             */
-/*   Updated: 2022/06/14 20:31:43 by vjean            ###   ########.fr       */
+/*   Updated: 2022/06/15 08:59:27 by vjean            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@
 void	ft_putchar(char c, int *len);
 void	ft_putnbr(int nb, int *len);
 void	ft_putstr(char *str, int *len);
-void	ft_puthex(unsigned short nb, int *len);
+void	ft_puthex(unsigned long nb, int *len);
 
 #endif
